+++ cli/bench/types.go
@@ -1,5 +1,12 @@
 package bench
 
+import (
+	"sync"
+
+	"github.com/FZambia/tarantool"
+	"github.com/tarantool/cartridge-cli/cli/context"
+)
+
 // Results describes set of benchmark results.
 type Results struct {
 	handledRequestsCount int     // Count of all executed requests.
@@ -7,4 +14,31 @@ type Results struct {
 	failedResultCount    int     // Count of failed request in all connections.
 	duration             float64 // Benchmark duration.
 	requestsPerSecond    int     // Cumber of requests per second - the main measured value.
+}
+
+// Request describes various types of requests.
+type Request struct {
+	operation             func(Request) // insertOperation, selectOperation or updateOperation.
+	ctx                   context.BenchCtx
+	tarantoolConnection   *tarantool.Connection
+	results               *Results
+	getRandomTupleCommand *string // for selectOperation or updateOperation.
+}
+
+// RequestsGenerator data structure for abstraction of a renewable heap of identical requests.
+type RequestsGenerator struct {
+	request Request // Request with specified operation.
+	count   int     // Count of requests.
+}
+
+// RequestsSequence data structure for abstraction for the constant issuance of new requests.
+type RequestsSequence struct {
+	requests []RequestsGenerator
+	// currentRequestIndex describes what type of request will be issued by the sequence.
+	currentRequestIndex int
+	// currentCounter describes how many requests of the same type
+	// are left to issue from RequestsPool.
+	currentCounter int
+	// findNewRequestsGeneratorMutex provides thread-safe search for new generator.
+	findNewRequestsGeneratorMutex sync.Mutex
 }